chore: refactored to have clean startup
